separate file convert* method
